feat: 0.0.15 图生图版本修改

+++ auiu-meta-support/meta-component-api/src/main/java/com/auiucloud/component/sd/domain/SdTxt2ImgParams.java
@@ -132,12 +132,12 @@ public class SdTxt2ImgParams implements Serializable {
     /**
      * 随机种子 -1
      */
-    private Integer seed = -1;
+    private Long seed = -1L;
 
     /**
      * 次级种子
      */
-    private Integer subseed = -1;
+    private Long subseed = -1L;
 
     /**
      * 次级种子强度
